requirements.txt
requests>=2.31.0
typing-extensions>=4.8.0
python-dateutil>=2.8.2
Flask==2.3.2
Flask-Session==0.5.0
Werkzeug<3.0.0